+++ HISTORY.md
@@ -1,3 +1,19 @@
+## >= 4.0.0
+
+* Some variable have been renamed:
+  * `clickCallback` -> `onPageChange`
+  * `initialSelected` -> `initialPage`
+  * `forceSelected` -> `forcePage`
+  * `pageNum` -> `pageCount`
+
+* `onClick` events have been moved on `<a>` tags (previously on `<li>`s).
+
+
+## >= 3.0.0
+
+`clickCallback` (`onPageChange`) isn't called during initialization anymore.
+
+
 ## >= 1.0.0
 
 HTML Structure:
@@ -13,6 +29,7 @@ HTML Structure:
   <li><a href="#"><span>»</span></a></li>
 </ul>
 ```
+
 
 ## <= 0.5.7